Implement job assignment and unassignment
This commit introduces functionality for assigning and unassigning companies to jobs.

Removed the JobAssignDto as the company email is retrieved from the user's token.
Updated the assignCompanyToJob controller method to retrieve the email from the token and pass it to the service.
Added a new unassignCompanyFromJob controller method to remove the company association from the job.
Implemented the assignCompany and unassignCompany methods in the jobs service.
Updated the jobs routes in the server to use the new methods and to use PUT method.
Updated the JobList component to include buttons for assigning and unassigning the job.
Updated the ListOfAllJobs and ListOfMyJobs components to show the assign/unassign buttons.
Added logic for checking if a company email is assigned or not before doing any assignment.
Added logic for checking if the unassigned company email is the same as the current company email.

These changes provide a more secure and convenient way to handle job assignments.